refactoring of magics and endpoint handler
